requirements.txt
flask==3.0.3
requests==2.32.3
vaderSentiment==3.3.2
python-dotenv==1.0.1


# Optional (enable only if you want ML summarization; heavy downloads)
# transformers==4.44.2
# torch==2.4.0